refactor(MeetingItemCard): MeetingItemCard 컴포넌트를 리팩토링해서 스와이프 상태 관련 로직은 useMeetingItemSwiperController 커스텀 훅으로 분리
