+++ eslint-suppressions.json
@@ -34,11 +34,6 @@
       "count": 1
     }
   },
-  "src/pages/meeting/list/MeetingCompleteSection.tsx": {
-    "@typescript-eslint/no-unused-vars": {
-      "count": 1
-    }
-  },
   "src/pages/meeting/list/MeetingIncompleteSection.tsx": {
     "@typescript-eslint/no-unused-vars": {
       "count": 1
